# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DriveTrack - Driving School Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /users/me/role:
    patch:
      operationId: updateMyRole
      tags: [users]
      summary: Set or update user role (admin only in prod; open for onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── Students ────────────────────────────────────────────────────────────────
...
  /viewer-links/request:
    post:
      operationId: requestViewerLink
      tags: [viewer]
      summary: Link to a student using a viewer code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewerLinkRequest"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /viewer/students/{id}/dashboard:
    get:
      operationId: getViewerStudentDashboard
      tags: [viewer]
      summary: Get sanitised student dashboard for a viewer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Viewer student dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewerStudentDashboard"

  # ─── Moderation ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, role, email]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, instructor, admin, school_admin, viewer, super_admin, unassigned]
        createdAt:
          type: string
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [student, instructor, admin, school_admin, viewer, super_admin]

    # ─── Students ──────────────────────────────────────────────────────────────
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── Schools ──────────────────────────────────────────────────────────────
...
    ViewerLinkRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: "Student viewer code (e.g. DRV-7KQ9X2)"
        relationshipType:
          type: string
          enum: [parent, guardian, mentor, support_worker, agency_case_worker, school_mentor, other]
    ViewerStudentSummary:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: integer
        fullName:
          type: string
        totalHours:
          type: ["number", "null"]
        headshotPath:
          type: ["string", "null"]
        noShowCount:
          type: integer
        attendanceReliabilityScore:
          type: ["integer", "null"]
        relationshipType:
          type: ["string", "null"]
        linkedAt:
          type: string
    ViewerStudentDashboard:
      type: object
      required: [student, recentAssessments, upcomingBookings]
      properties:
        student:
          $ref: "#/components/schemas/ViewerStudentSummary"
        recentAssessments:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              lessonDate:
                type: string
              durationMinutes:
                type: integer
              pedalOperator:
                type: string
              focusAreasNext:
                type: ["string", "null"]
              totalHoursThisLesson:
                type: ["number", "null"]
        upcomingBookings:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              scheduledAt:
                type: string
              durationMinutes:
                type: integer
              status:
                type: string
              pickupAddress:
                type: ["string", "null"]
        link:
          type: object
          properties:
            relationshipType:
              type: ["string", "null"]
            linkedAt:
              type: string

    # ─── Moderation (detail) ──────────────────────────────────────────────────
...